PreferenceViewModelクラスに入れ子になっていたPasswordクラスを分離した。#2
